gnome-source: fix pull-source.sh layout bug and add version pinning
Fix the clone-to-root bug and pin GNOME module versions so
pull-all-source.sh works cleanly from a fresh checkout.

pull-source.sh (glib, gdk-pixbuf, gtk, mutter, gnome-shell, pango, and now
gvfs, orca):
- Clone/extract directly into <module>/source/ via an atomic swap
  (source.new -> mv), instead of dumping the upstream tree into the module
  root and cp -a'ing over existing files. This eliminates the
  'cp: cannot overwrite non-directory' failures and makes re-runs idempotent.
- Accept a pinned ref via <MODULE>_REF (git tag/branch/commit) and
  <MODULE>_URL override; record the resolved commit in <module>.source-ref.
- gvfs/orca are dual-mode: git clone by default (works without
  download.gnome.org), or a verified release tarball when <MODULE>_VERSION
  is set.

pull-all-source.sh:
- Load version pins from gnome-source/GNOME_VERSIONS before each fetch
  (env overrides the file); add FORCE=1 to re-fetch present modules; drop the
  now-unneeded relocate-root workaround.

GNOME_VERSIONS: pin the git-based modules to verified release tags
(glib 2.80.4, gdk-pixbuf 2.42.12, gtk 4.14.5, mutter 46.4, gnome-shell 46.4,
gvfs 1.56.1, orca 50.2, pango 1.54.0).

Re-fetched all seven git modules to their pinned versions (previously tip of
main). Docs updated in build/README.md.

+++ gnome-source/glib/source/glib/deprecated/gthread-deprecated.c
@@ -135,7 +135,7 @@ GThreadFunctions g_thread_functions_for_glib_use =
 static guint64
 gettime (void)
 {
-  return (guint64) g_get_monotonic_time () * 1000;
+  return g_get_monotonic_time () * 1000;
 }
 
 guint64 (*g_thread_gettime) (void) = gettime;
@@ -158,17 +158,15 @@ gboolean         g_threads_got_initialized = TRUE;
  * Since version 2.32, GLib does not support custom thread implementations
  * anymore and the @vtable parameter is ignored and you should pass %NULL.
  *
- * ::: note
- *     g_thread_init() must not be called directly or indirectly in a
- *     callback from GLib. Also no mutexes may be currently locked
- *     while calling g_thread_init().
+ * <note><para>g_thread_init() must not be called directly or indirectly
+ * in a callback from GLib. Also no mutexes may be currently locked while
+ * calling g_thread_init().</para></note>
  *
- * ::: note
- *     To use g_thread_init() in your program, you have to link with
- *     the libraries that the command `pkg-config --libs gthread-2.0`
- *     outputs. This is not the case for all the other thread-related
- *     functions of GLib. Those can be used without having to link
- *     with the thread libraries.
+ * <note><para>To use g_thread_init() in your program, you have to link
+ * with the libraries that the command <command>pkg-config --libs
+ * gthread-2.0</command> outputs. This is not the case for all the
+ * other thread-related functions of GLib. Those can be used without
+ * having to link with the thread libraries.</para></note>
  *
  * Deprecated:2.32: This function is no longer necessary. The GLib
  *     threading system is automatically initialized at the start
@@ -803,8 +801,8 @@ guint
 g_static_rec_mutex_unlock_full (GStaticRecMutex *mutex)
 {
   GRecMutex *rm;
-  guint depth;
-  guint i;
+  gint depth;
+  gint i;
 
   rm = g_static_rec_mutex_get_rec_mutex_impl (mutex);
 
@@ -851,7 +849,7 @@ g_static_rec_mutex_free (GStaticRecMutex *mutex)
 /**
  * GStaticRWLock: (skip):
  *
- * The [struct@StaticRWLock] struct represents a read-write lock. A read-write
+ * The #GStaticRWLock struct represents a read-write lock. A read-write
  * lock can be used for protecting data that some portions of code only
  * read from, while others also write. In such situations it is
  * desirable that several readers can read at once, whereas of course
@@ -895,8 +893,8 @@ g_static_rec_mutex_free (GStaticRecMutex *mutex)
  * ]|
  *
  * This example shows an array which can be accessed by many readers
- * (the `my_array_get()` function) simultaneously, whereas the writers
- * (the `my_array_set()` function) will only be allowed once at a time
+ * (the my_array_get() function) simultaneously, whereas the writers
+ * (the my_array_set() function) will only be allowed once at a time
  * and only if no readers currently access the array. This is because
  * of the potentially dangerous resizing of the array. Using these
  * functions is fully multi-thread safe now.
@@ -908,23 +906,23 @@ g_static_rec_mutex_free (GStaticRecMutex *mutex)
  * to finish their operation. As soon as the last reader unlocks the
  * data, the writer will lock it.
  *
- * Even though [struct@StaticRWLock] is not opaque, it should only be used
+ * Even though #GStaticRWLock is not opaque, it should only be used
  * with the following functions.
  *
- * All of the `g_static_rw_lock_*` functions can be used even if
- * [func@Thread.init] has not been called. Then they do nothing, apart
- * from `g_static_rw_lock_*_trylock`, which does nothing but returning true.
+ * All of the g_static_rw_lock_* functions can be used even if
+ * g_thread_init() has not been called. Then they do nothing, apart
+ * from g_static_rw_lock_*_trylock, which does nothing but returning %TRUE.
  *
  * A read-write lock has a higher overhead than a mutex. For example, both
- * [method@StaticRWLock.reader_lock] and [method@StaticRWLock.reader_unlock] have
- * to lock and unlock a [struct@StaticMutex], so it takes at least twice the time
- * to lock and unlock a [struct@StaticRWLock] that it does to lock and unlock a
- * [struct@StaticMutex]. So only data structures that are accessed by multiple
+ * g_static_rw_lock_reader_lock() and g_static_rw_lock_reader_unlock() have
+ * to lock and unlock a #GStaticMutex, so it takes at least twice the time
+ * to lock and unlock a #GStaticRWLock that it does to lock and unlock a
+ * #GStaticMutex. So only data structures that are accessed by multiple
  * readers, and which keep the lock for a considerable time justify a
- * [struct@StaticRWLock]. The above example most probably would fare better with a
- * [struct@StaticMutex].
+ * #GStaticRWLock. The above example most probably would fare better with a
+ * #GStaticMutex.
  *
- * Deprecated: 2.32: Use a [struct@RWLock] instead
+ * Deprecated: 2.32: Use a #GRWLock instead
  **/
 
 /**
@@ -1196,7 +1194,7 @@ g_static_rw_lock_free (GStaticRWLock* lock)
 /* GPrivate {{{1 ------------------------------------------------------ */
 
 /**
- * g_private_new: (skip) (constructor) (not method):
+ * g_private_new: (skip):
  * @notify: a #GDestroyNotify
  *
  * Creates a new #GPrivate.
@@ -1204,7 +1202,7 @@ g_static_rw_lock_free (GStaticRWLock* lock)
  * Deprecated:2.32: dynamic allocation of #GPrivate is a bad idea.  Use
  *                  static storage and G_PRIVATE_INIT() instead.
  *
- * Returns: (transfer full): a newly allocated #GPrivate (which can never be destroyed)
+ * Returns: a newly allocated #GPrivate (which can never be destroyed)
  */
 GPrivate *
 g_private_new (GDestroyNotify notify)
@@ -1443,11 +1441,11 @@ g_static_private_free (GStaticPrivate *private_key)
 /* GMutex {{{1 ------------------------------------------------------ */
 
 /**
- * g_mutex_new: (skip) (constructor) (not method):
+ * g_mutex_new: (skip):
  *
  * Allocates and initializes a new #GMutex.
  *
- * Returns: (transfer full): a newly allocated #GMutex. Use g_mutex_free() to free
+ * Returns: a newly allocated #GMutex. Use g_mutex_free() to free
  *
  * Deprecated: 2.32: GMutex can now be statically allocated, or embedded
  * in structures and initialised with g_mutex_init().
@@ -1485,11 +1483,11 @@ g_mutex_free (GMutex *mutex)
 /* GCond {{{1 ------------------------------------------------------ */
 
 /**
- * g_cond_new: (skip) (constructor) (not method):
+ * g_cond_new: (skip):
  *
  * Allocates and initializes a new #GCond.
  *
- * Returns: (transfer full): a newly allocated #GCond. Free with g_cond_free()
+ * Returns: a newly allocated #GCond. Free with g_cond_free()
  *
  * Deprecated: 2.32: GCond can now be statically allocated, or embedded
  * in structures and initialised with g_cond_init().
