+++ gnome-source/glib/source/gobject/gbinding.c
@@ -256,8 +256,8 @@ struct _GBinding
 
   GBindingFlags flags;
 
-  unsigned long source_notify; /* LOCK: unbind_lock */
-  unsigned long target_notify; /* LOCK: unbind_lock */
+  guint source_notify; /* LOCK: unbind_lock */
+  guint target_notify; /* LOCK: unbind_lock */
   gboolean target_weak_notify_installed; /* LOCK: unbind_lock */
 
   /* a guard, to avoid loops */
@@ -910,7 +910,7 @@ g_binding_class_init (GBindingClass *klass)
    * The name of the property of #GBinding:source that should be used
    * as the source of the binding.
    *
-   * This should be in [canonical form][class@GObject.ParamSpec#parameter-names] to get the
+   * This should be in [canonical form][canonical-parameter-names] to get the
    * best performance.
    *
    * Since: 2.26
@@ -927,7 +927,7 @@ g_binding_class_init (GBindingClass *klass)
    * The name of the property of #GBinding:target that should be used
    * as the target of the binding.
    *
-   * This should be in [canonical form][class@GObject.ParamSpec#parameter-names] to get the
+   * This should be in [canonical form][canonical-parameter-names] to get the
    * best performance.
    *
    * Since: 2.26
@@ -1250,7 +1250,7 @@ g_object_bind_property_full (gpointer               source,
   if ((flags & G_BINDING_INVERT_BOOLEAN) &&
       (transform_to != NULL || transform_from != NULL))
     {
-      flags &= (unsigned) ~G_BINDING_INVERT_BOOLEAN;
+      flags &= ~G_BINDING_INVERT_BOOLEAN;
     }
 
   pspec = g_object_class_find_property (G_OBJECT_GET_CLASS (source), source_property);
